% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetHighResolutionImage.R
\name{GetHighResolutionImage}
\alias{GetHighResolutionImage}
\title{Get image limited to an area of interest}
\source{
\url{https://api.spectator.earth/#high-resolution-image}
}
\usage{
GetHighResolutionImage(
  aoi,
  id,
  bands,
  width,
  height,
  file = "image.jpg",
  api_key = Sys.getenv("spectator_earth_api_key")
)
}
\arguments{
\item{aoi}{'\code{sf}' (or '\code{Spatial*}') object defining the area of interest.
Can be of any geometry as only the bounding box is used.}

\item{id}{integer, \code{id} of the image from the \code{\link[spectator]{SearchImages}} result}

\item{bands}{integer vector of length 1 or 3 indicating the spectral bands to use for creating the image
(typically the bands corresponding to Red, Green and Blue)}

\item{width}{integer indicating the width of the image (in pixels)}

\item{height}{integer indicating the height of the image (in pixels)}

\item{file}{character indicating the name of the image file to create. Default: 'image.jpg'}

\item{api_key}{character containing your API key. Default: \code{Sys.getenv("spectator_earth_api_key")}}
}
\value{
The name of the image file is returned invisibly.
}
\description{
Provides you with a high resolution image for the area within the acquisition boundaries,
if you're not interested in downloading the whole image.
}
\details{
As a side effect, the image file is written to the provided path.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[spectator]{SearchImages}}
}
