% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetImageryFilesList.R
\name{GetImageryFilesList}
\alias{GetImageryFilesList}
\title{FUNCTION_TITLE}
\source{
\url{http://somewhere.important.com/}
}
\usage{
GetImageryFilesList(id, api_key = Sys.getenv("spectator_earth_api_key"))
}
\arguments{
\item{id}{PARAM_DESCRIPTION}

\item{api_key}{PARAM_DESCRIPTION, Default: Sys.getenv("spectator_earth_api_key")}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[httr]{GET}},\code{\link[httr]{content}}
}
