% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetOverpasses.R
\name{GetOverpasses}
\alias{GetOverpasses}
\title{Get overpasses}
\source{
\url{http://somewhere.important.com/}
}
\usage{
GetOverpasses(
  aoi,
  satellites = NULL,
  days_before = 0,
  days_after = 7,
  acquisitions = TRUE,
  api_key = Sys.getenv("spectator_earth_api_key")
)
}
\arguments{
\item{aoi}{PARAM_DESCRIPTION}

\item{satellites}{PARAM_DESCRIPTION. Default: NULL}

\item{days_before}{PARAM_DESCRIPTION. Default: 0}

\item{days_after}{PARAM_DESCRIPTION. Default: 7}

\item{acquisitions}{PARAM_DESCRIPTION. Default: TRUE}

\item{api_key}{PARAM_DESCRIPTION. Default: Sys.getenv("spectator_earth_api_key")}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Get overpasses
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[sf]{st_as_sf}}, \code{\link[sf]{st_bbox}}
\code{\link[httr]{GET}}, \code{\link[httr]{content}}
\code{\link[sp]{c("SpatialPolygons", "polygons")}}, \code{\link[sp]{SpatialPolygons}}
}
