% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSatelliteOverpasses.R
\name{GetSatelliteOverpasses}
\alias{GetSatelliteOverpasses}
\title{Get satellite overpasses}
\source{
\url{http://somewhere.important.com/}
}
\usage{
GetSatelliteOverpasses(
  aoi,
  satellites = NULL,
  days_before = 0,
  days_after = 7,
  acquisitions = TRUE,
  api_key = Sys.getenv("api_key")
)
}
\arguments{
\item{aoi}{PARAM_DESCRIPTION}

\item{satellites}{PARAM_DESCRIPTION. Default: NULL}

\item{days_before}{PARAM_DESCRIPTION. Default: 0}

\item{days_after}{PARAM_DESCRIPTION. Default: 7}

\item{acquisitions}{PARAM_DESCRIPTION. Default: TRUE}

\item{api_key}{PARAM_DESCRIPTION. Default: Sys.getenv("api_key")}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Get satellite overpasses
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[sp]{bbox-methods}}, \code{\link[sp]{c("SpatialPolygons", "polygons")}}, \code{\link[sp]{SpatialPolygons}}, \code{\link[sp]{CRS-class}}, \code{\link[sp]{coordinates}}, \code{\link[sp]{is.projected}}, \code{\link[sp]{SpatialPoints}}
\code{\link[httr]{GET}}, \code{\link[httr]{content}}
}
