% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSatelliteTrajectory.R
\name{GetSatelliteTrajectory}
\alias{GetSatelliteTrajectory}
\title{Get satellite trajectory}
\source{
\url{http://somewhere.important.com/}
}
\usage{
GetSatelliteTrajectory(satellite)
}
\arguments{
\item{satellite}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Get satellite trajectory
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[httr]{GET}}, \code{\link[httr]{content}}
\code{\link[geojsonio]{geojson_sp}}
}
