% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetAcquisitionPlan.R
\name{GetAcquisitionPlan}
\alias{GetAcquisitionPlan}
\title{Get acquisition plan}
\source{
\url{http://somewhere.important.com/}
}
\usage{
GetAcquisitionPlan(satellites = NULL, date = NULL)
}
\arguments{
\item{satellites}{PARAM_DESCRIPTION. Default: NULL}

\item{date}{PARAM_DESCRIPTION. Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Get acquisition plan
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[httr]{GET}}, \code{\link[httr]{content}}
\code{\link[geojsonio]{geojson_sp}}
}
