% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetAllSatellites.R
\name{GetAllSatellites}
\alias{GetAllSatellites}
\title{Gets all referenced satellites info}
\source{
\url{https://api.spectator.earth/#satellites}
}
\usage{
GetAllSatellites(positions = TRUE)
}
\arguments{
\item{positions}{logical indicating if the current position should be included. Default: TRUE}
}
\value{
If \code{positions} is \code{FALSE}, a data frame with following attributes:
\describe{
\item{\code{id}}{integer identifier}
\item{\code{name}}{character satellite name}
\item{\code{norad_id}}{integer satellite catalog number}
\item{\code{sensors}}{character type of sensors available on the satellite (SAR or Optical)}
\item{\code{open}}{logical whether the data produced by the satellite is freely accessible}
\item{\code{platform}}{character platform name}
}
If \code{positions} is \code{TRUE}, object of class '\code{sf}' with '\code{POINT}' geometry type,
with the same attributes as above.
}
\description{
Gets the information about all the satellites known in the Spectator Earth database,
and possibly their current positions.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[spectator]{GetSatellite}}
}
