% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetAllSatellites.R
\name{GetAllSatellites}
\alias{GetAllSatellites}
\title{Get all satellites}
\source{
\url{http://somewhere.important.com/}
}
\usage{
GetAllSatellites(satellites = NULL, date = NULL)
}
\arguments{
\item{satellites}{PARAM_DESCRIPTION. Default: NULL}

\item{date}{PARAM_DESCRIPTION. Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Get all satellites
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[httr]{GET}}, \code{\link[httr]{content}}
}
