% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetAllSatellites.R
\name{GetAllSatellites}
\alias{GetAllSatellites}
\title{Get all satellites}
\source{
\url{http://somewhere.important.com/}
}
\usage{
GetAllSatellites(positions = TRUE)
}
\arguments{
\item{positions}{PARAM_DESCRIPTION. Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Get all satellites
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[httr]{GET}}, \code{\link[httr]{content}}
\code{\link[geojsonsf]{geojson_sf}}
}
