% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetHighResolutionImage.R
\name{GetHighResolutionImage}
\alias{GetHighResolutionImage}
\title{FUNCTION_TITLE}
\source{
\url{http://somewhere.important.com/}
}
\usage{
GetHighResolutionImage(
  aoi,
  id,
  bands,
  width,
  height,
  file = "image.jpg",
  api_key = Sys.getenv("spectator_earth_api_key")
)
}
\arguments{
\item{aoi}{PARAM_DESCRIPTION}

\item{id}{PARAM_DESCRIPTION}

\item{bands}{PARAM_DESCRIPTION}

\item{width}{PARAM_DESCRIPTION}

\item{height}{PARAM_DESCRIPTION}

\item{file}{PARAM_DESCRIPTION, Default: 'image.jpg'}

\item{api_key}{PARAM_DESCRIPTION, Default: Sys.getenv("spectator_earth_api_key")}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[sf]{st_as_sf}},\code{\link[sf]{st_bbox}}
\code{\link[httr]{GET}},\code{\link[httr]{content}}
}
