% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSatellite.R
\name{GetSatellite}
\alias{GetSatellite}
\title{Get satellite}
\source{
\url{http://somewhere.important.com/}
}
\usage{
GetSatellite(satellite, positions = TRUE)
}
\arguments{
\item{satellite}{PARAM_DESCRIPTION}

\item{positions}{PARAM_DESCRIPTION. Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Get satellite
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[httr]{GET}}, \code{\link[httr]{content}}
\code{\link[geojsonio]{geojson_sf}}
}
