% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetTrajectory.R
\name{GetTrajectory}
\alias{GetTrajectory}
\title{Gets a satellite trajectory}
\source{
\url{https://api.spectator.earth/#trajectories}
}
\usage{
GetTrajectory(satellite)
}
\arguments{
\item{satellite}{character name of the satellite for which to retrieve the trajectory.
The satellite name is not case sensitive, and can be abbreviated as long as an unambiguous match can be obtained.
Only one satellite can be queried at a time.}
}
\value{
Object of class '\code{sf}' with '\code{LINESTRING}' geometry type
}
\description{
Gets the current trajectory for the specified satellite.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
