% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetTrajectory.R
\name{GetTrajectory}
\alias{GetTrajectory}
\title{Get trajectory}
\source{
\url{http://somewhere.important.com/}
}
\usage{
GetTrajectory(satellite)
}
\arguments{
\item{satellite}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Get trajectory
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[httr]{GET}}, \code{\link[httr]{content}}
\code{\link[geojsonsf]{geojson_sf}}
}
