% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SearchImages.R
\name{SearchImages}
\alias{SearchImages}
\title{FUNCTION_TITLE}
\source{
\url{http://somewhere.important.com/}
}
\usage{
SearchImages(
  aoi,
  satellites = NULL,
  from = NULL,
  date_to = NULL,
  footprint = FALSE,
  api_key = Sys.getenv("spectator_earth_api_key")
)
}
\arguments{
\item{aoi}{PARAM_DESCRIPTION}

\item{satellites}{PARAM_DESCRIPTION, Default: NULL}

\item{from}{PARAM_DESCRIPTION, Default: NULL}

\item{date_to}{PARAM_DESCRIPTION, Default: NULL}

\item{footprint}{PARAM_DESCRIPTION, Default: FALSE}

\item{api_key}{PARAM_DESCRIPTION, Default: Sys.getenv("spectator_earth_api_key")}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[sf]{st_as_sf}},\code{\link[sf]{st_bbox}},\code{\link[sf]{st}},\code{\link[sf]{sfc}},\code{\link[sf]{sf}}
\code{\link[httr]{GET}},\code{\link[httr]{content}}
}
